refactor(IT-Wallet): [SIW-4365] extract CIE PIN flow
