+++ apps/main-app/ts/features/itwallet/machine/eid/machine.ts
@@ -1,261 +1,13 @@
-import _ from "lodash";
-import {
-  and,
-  assertEvent,
-  assign,
-  fromCallback,
-  fromPromise,
-  not,
-  or,
-  raise,
-  setup
-} from "xstate";
+import { and, assertEvent, assign, not, or, raise } from "xstate";
 
 import { assert } from "../../../../utils/assert.ts";
-import { trackItWalletIntroScreen } from "../../analytics";
-import {
-  CredentialAccessToken,
-  WalletInstanceAttestations
-} from "../../common/utils/itwTypesUtils";
-import { isMrtdPoPChallengeRequired } from "../../common/utils/mrtdUrl";
 import { ItwTags } from "../tags";
-import { itwCredentialUpgradeMachine } from "../upgrade/machine.ts";
-import {
-  CreateWalletInstanceActorParams,
-  GetWalletAttestationActorParams,
-  InitMrtdPoPChallengeActorParams,
-  ObtainStatusListActorInput,
-  ObtainStatusListActorOutput,
-  RequestAccessTokenActorParams,
-  RequestEidActorOutput,
-  type RequestEidActorParams,
-  StartAuthFlowActorParams,
-  StoreEidCredentialActorParams,
-  ValidateMrtdPoPChallengeActorParams,
-  WithItwVersion
-} from "./actors";
-import {
-  AuthenticationContext,
-  CieContext,
-  Context,
-  InitialContext,
-  MrtdPoPContext
-} from "./context";
-import { EidIssuanceEvents } from "./events";
-import { IssuanceFailureType, mapEventToFailure } from "./failure";
+import { InitialContext } from "./context";
+import { IssuanceFailureType } from "./failure";
+import { itwEidIssuanceMachineSetup } from "./setup";
+import { ciePinState } from "./states/ciePin";
 
-const notImplemented = () => {
-  throw new Error("Not implemented");
-};
-
-export const itwEidIssuanceMachine = setup({
-  types: {
-    context: {} as Context,
-    events: {} as EidIssuanceEvents
-  },
-  actions: {
-    onInit: notImplemented,
-
-    /**
-     * Navigation
-     */
-
-    navigateToTosScreen: notImplemented,
-    navigateToIpzsPrivacyScreen: notImplemented,
-    navigateToIdentificationScreen: notImplemented,
-    navigateToIdpSelectionScreen: notImplemented,
-    navigateToSpidLoginScreen: notImplemented,
-    navigateToCieIdLoginScreen: notImplemented,
-    navigateToEidPreviewScreen: notImplemented,
-    navigateToSuccessScreen: notImplemented,
-    navigateToFailureScreen: notImplemented,
-    navigateToWallet: notImplemented,
-    navigateToCredentialCatalog: notImplemented,
-    navigateToCieNfcPreparationScreen: notImplemented,
-    navigateToCiePinPreparationScreen: notImplemented,
-    navigateToCieCardPreparationScreen: notImplemented,
-    navigateToCieCanPreparationScreen: notImplemented,
-    navigateToCiePinScreen: notImplemented,
-    navigateToCieAuthenticationScreen: notImplemented,
-    navigateToNfcInstructionsScreen: notImplemented,
-    navigateToWalletRevocationScreen: notImplemented,
-    navigateToCieWarningScreen: notImplemented,
-    navigateToCieCanScreen: notImplemented,
-    navigateToCieInternalAuthAndMrtdScreen: notImplemented,
-    closeIssuance: notImplemented,
-
-    /**
-     * Store update
-     */
-
-    storeIntegrityKeyTag: notImplemented,
-    cleanupIntegrityKeyTag: notImplemented,
-    storeWalletInstanceAttestation: notImplemented,
-    storeAuthLevel: notImplemented,
-    storeWalletActivationFeedbackBannerData: notImplemented,
-    storeCredentialUpgradeFailures: notImplemented,
-    handleSessionExpired: notImplemented,
-    resetWalletInstance: notImplemented,
-    refreshCredentialsCatalogue: notImplemented,
-
-    /**
-     * Analytics
-     */
-
-    trackWalletInstanceCreation: notImplemented,
-    trackWalletInstanceRevocation: notImplemented,
-    trackIdentificationMethodSelected: notImplemented,
-    trackItwIdAuthenticationCompleted: notImplemented,
-    trackItwIdVerifiedDocument: notImplemented,
-    /**
-     * Context manipulation
-     */
-
-    setCieIdIdentificationL2: assign(() => ({
-      identification: {
-        mode: "cieId",
-        level: "L2"
-      } as const
-    })),
-
-    /**
-     * Updates the CieID identification level to L3 when IPZS confirms native L3
-     * authentication (i.e. challenge_info is absent in the callback URL, meaning
-     * no MRTD PoP is required because the CieID app already authenticated at L3).
-     */
-    updateCieIdIdentificationLevel: assign(({ context, event }) => {
-      assertEvent(event, "user-identification-completed");
-      if (
-        context.identification?.mode !== "cieId" ||
-        context.level !== "l3" ||
-        isMrtdPoPChallengeRequired(event.authRedirectUrl)
-      ) {
-        return {};
-      }
-      return { identification: { mode: "cieId", level: "L3" } as const };
-    }),
-    setFailure: assign(({ event }) => ({ failure: mapEventToFailure(event) })),
-    /**
-     * Save the final redirect url in the machine context for later reuse.
-     * This action is the same for the three identification methods.
-     */
-    completeUserIdentification: assign(({ context, event }) => {
-      assertEvent(event, "user-identification-completed");
-      assert(
-        context.authenticationContext,
-        "authenticationContext must be defined when completing auth flow"
-      );
-      return {
-        authenticationContext: {
-          ...context.authenticationContext,
-          callbackUrl: event.authRedirectUrl
-        }
-      };
-    }),
-    completeMrtdPoP: assign(({ context, event }) => {
-      assertEvent(event, "mrtd-pop-verification-completed");
-      assert(
-        context.authenticationContext,
-        "authenticationContext must be defined when completing auth flow"
-      );
-      return {
-        authenticationContext: {
-          ...context.authenticationContext,
-          callbackUrl: event.authRedirectUrl
-        }
-      };
-    }),
-    trackIntroScreen: ({ context }) => {
-      trackItWalletIntroScreen(context.level === "l3" ? "L3" : "L2");
-    }
-  },
-  actors: {
-    getCieStatus: fromPromise<CieContext>(notImplemented),
-    verifyTrustFederation: fromPromise<void, WithItwVersion>(notImplemented),
-
-    /**
-     * WI actors
-     */
-
-    createWalletInstance: fromPromise<string, CreateWalletInstanceActorParams>(
-      notImplemented
-    ),
-    revokeWalletInstance: fromPromise<void, WithItwVersion>(notImplemented),
-    getWalletAttestation: fromPromise<
-      WalletInstanceAttestations,
-      GetWalletAttestationActorParams
-    >(notImplemented),
-
-    /**
-     * Primary authentication actors
-     */
-
-    startAuthFlow: fromPromise<AuthenticationContext, StartAuthFlowActorParams>(
-      notImplemented
-    ),
-
-    /**
-     * MRTD PoP Challenge actors
-     */
-
-    initMrtdPoPChallenge: fromPromise<
-      MrtdPoPContext,
-      InitMrtdPoPChallengeActorParams
-    >(notImplemented),
-    validateMrtdPoPChallenge: fromPromise<
-      string,
-      ValidateMrtdPoPChallengeActorParams
-    >(notImplemented),
-
-    /**
-     * PID issuance actors
-     */
-
-    requestAccessToken: fromPromise<
-      CredentialAccessToken,
-      RequestAccessTokenActorParams
-    >(notImplemented),
-    requestEid: fromPromise<RequestEidActorOutput, RequestEidActorParams>(
-      notImplemented
-    ),
-    obtainStatusList: fromPromise<
-      ObtainStatusListActorOutput,
-      ObtainStatusListActorInput
-    >(notImplemented),
-    storeEidCredential: fromPromise<void, StoreEidCredentialActorParams>(
-      notImplemented
-    ),
-    waitForSessionRefresh: fromCallback(notImplemented),
-
-    /**
-     * Credential upgrade actors
-     */
-
-    credentialUpgradeMachine: itwCredentialUpgradeMachine
-  },
-  guards: {
-    issuedEidMatchesAuthenticatedUser: notImplemented,
-    isSessionExpired: notImplemented,
-    isOperationAborted: notImplemented,
-    hasIntegrityKeyTag: ({ context }) => context.integrityKeyTag !== undefined,
-    hasValidWalletInstanceAttestation: notImplemented,
-    hasCredentialsToUpgrade: ({ context }) =>
-      context.credentialsToUpgrade.length > 0,
-    isNFCEnabled: ({ context }) => context.cieContext?.isNFCEnabled || false,
-    isReissuance: ({ context }) => context.mode === "reissuance",
-    isUpgrade: ({ context }) => context.mode === "upgrade",
-    isL2Fallback: ({ context }) => context.level === "l2-fallback",
-    isL3FeaturesEnabled: ({ context }) => context.level === "l3",
-    requiresMrtdVerification: ({ context }) =>
-      // MRTD PoP verification is required for SPID and CieID identification modes
-      // when issuing an L3 PID and the PID Provider signals a challenge via `challenge_info`.
-      context.level === "l3" &&
-      context.identification?.mode !== "ciePin" &&
-      context.authenticationContext !== undefined &&
-      isMrtdPoPChallengeRequired(context.authenticationContext.callbackUrl),
-    isWalletValid: notImplemented
-  }
-}).createMachine({
+export const itwEidIssuanceMachine = itwEidIssuanceMachineSetup.createMachine({
   id: "itwEidIssuanceMachine",
   context: { ...InitialContext },
   initial: "Idle",
@@ -741,191 +493,7 @@ export const itwEidIssuanceMachine = setup({
             target: "#itwEidIssuanceMachine.UserIdentification.Completed"
           }
         },
-        CiePin: {
-          description:
-            "This state handles the entire CIE + pin identification flow",
-          initial: "PreparationPin",
-          states: {
-            PreparationPin: {
-              description:
-                "This state handles the CIE PIN preparation screen, where the user is informed about the CIE PIN",
-              entry: "navigateToCiePinPreparationScreen",
-              on: {
-                next: [
-                  {
-                    guard: "isNFCEnabled",
-                    target: "InsertingCardPin"
-                  },
-                  {
-                    target: "RequestingNfcActivation"
-                  }
-                ],
-                "go-to-cie-warning": {
-                  target: "CieWarning.PreparationPin"
-                },
-                back: {
-                  target: "#itwEidIssuanceMachine.UserIdentification"
-                },
-                close: {
-                  actions: "closeIssuance"
-                }
-              }
-            },
-            RequestingNfcActivation: {
-              entry: "navigateToNfcInstructionsScreen",
-              on: {
-                "nfc-enabled": {
-                  actions: assign(({ context }) => ({
-                    cieContext: _.merge(context.cieContext, {
-                      isNFCEnabled: true
-                    })
-                  })),
-                  target: "InsertingCardPin"
-                },
-                back: {
-                  target: "#itwEidIssuanceMachine.UserIdentification"
-                }
-              }
-            },
-            InsertingCardPin: {
-              entry: [
-                assign(() => ({ authenticationContext: undefined })), // Reset the authentication context, otherwise retries will use stale data
-                "navigateToCiePinScreen"
-              ],
-              on: {
-                "cie-pin-entered": {
-                  target: "PreparationCie",
-                  actions: assign(({ event }) => ({
-                    identification: {
-                      mode: "ciePin",
-                      level: "L3",
-                      pin: event.pin
-                    }
-                  }))
-                },
-                back: {
-                  target: "PreparationPin"
-                }
-              }
-            },
-            PreparationCie: {
-              description:
-                "This state handles the CIE preparation screen, where the user is informed about the CIE card",
-              entry: "navigateToCieNfcPreparationScreen",
-              on: {
-                next: {
-                  target: "StartingCieAuthFlow"
-                },
-                "go-to-cie-warning": {
-                  target: "CieWarning.PreparationCie"
-                },
-                back: {
-                  target: "PreparationPin"
-                },
-                close: {
-                  actions: "closeIssuance"
-                }
-              }
-            },
-            StartingCieAuthFlow: {
-              description:
-                "Start the preliminary phase of the CIE identification flow.",
-              tags: [ItwTags.Loading],
-              entry: "navigateToCieAuthenticationScreen",
-              invoke: {
-                src: "startAuthFlow",
-                input: ({ context }) => ({
-                  itwVersion: context.itwVersion,
-                  walletInstanceAttestation:
-                    context.walletInstanceAttestation?.jwt,
-                  identification: context.identification,
-                  withMRTDPoP: false
-                }),
-                onDone: {
-                  actions: assign(({ event }) => ({
-                    authenticationContext: event.output
-                  })),
-                  target: "ReadingCieCard"
-                },
-                onError: {
-                  actions: "setFailure",
-                  target: "#itwEidIssuanceMachine.Failure"
-                }
-              },
-              on: {
-                back: {
-                  target: "PreparationCie"
-                }
-              }
-            },
-            ReadingCieCard: {
-              description:
-                "Read the CIE card and get back a url to continue the PID issuing flow. This state also handles errors when reading the card.",
-              on: {
-                "user-identification-completed": {
-                  target: "#itwEidIssuanceMachine.UserIdentification.Completed",
-                  actions: ["completeUserIdentification", "storeAuthLevel"]
-                },
-                close: {
-                  target: "#itwEidIssuanceMachine.UserIdentification"
-                },
-                back: {
-                  target: "PreparationCie"
-                }
-              }
-            },
-            CieWarning: {
-              description: "Navigates to and handles the CIE warning screen.",
-              entry: "navigateToCieWarningScreen",
-              initial: "Identification",
-              states: {
-                Identification: {
-                  on: {
-                    back: "#itwEidIssuanceMachine.UserIdentification.Identification",
-                    close: {
-                      target: "#itwEidIssuanceMachine.Idle",
-                      actions: "closeIssuance"
-                    }
-                  }
-                },
-                PreparationCie: {
-                  on: {
-                    back: "#itwEidIssuanceMachine.UserIdentification.CiePin.PreparationCie"
-                  }
-                },
-                PreparationPin: {
-                  on: {
-                    back: "#itwEidIssuanceMachine.UserIdentification.CiePin.PreparationPin"
-                  }
-                }
-              },
-              on: {
-                close: {
-                  actions: "closeIssuance"
-                }
-              }
-            }
-          },
-          on: {
-            "select-identification-mode": [
-              {
-                guard: ({ event }) => event.mode === "spid",
-                target: "#itwEidIssuanceMachine.UserIdentification.Spid"
-              },
-              {
-                guard: ({ event }) => event.mode === "cieId",
-                actions: [
-                  "trackIdentificationMethodSelected",
-                  "setCieIdIdentificationL2"
-                ],
-                target: "#itwEidIssuanceMachine.UserIdentification.CieID"
-              }
-            ]
-          },
-          onDone: {
-            target: "#itwEidIssuanceMachine.UserIdentification.Completed"
-          }
-        },
+        CiePin: ciePinState,
         Completed: {
           type: "final"
         }
